rename remaining getValidNodesFilenames -> gatherFileNodes
