+++ packages/source-filesystem/src/utils/gatherFileNodes.ts
