+++ packages/source-filesystem/src/utils/tests/snapshots/gatherFileNodes.test.ts.md
@@ -1,6 +1,6 @@
-# Snapshot report for `packages/source-filesystem/src/utils/tests/getValidNodesFilenames.test.ts`
+# Snapshot report for `packages/source-filesystem/src/utils/tests/gatherFileNodes.test.ts`
 
-The actual snapshot is saved in `getValidNodesFilenames.test.ts.snap`.
+The actual snapshot is saved in `gatherFileNodes.test.ts.snap`.
 
 Generated by [AVA](https://avajs.dev).
